BackEnd/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
mariadb==1.1.11
mysqlclient==2.2.6
packaging==24.2
sqlparse==0.5.3
tzdata==2024.2
django-cors-headers